requirements.txt
python-dotenv
pandas
numpy
scikit-learn
spacy>=3.5
beautifulsoup4
requests
PyPDF2
biopython
networkx
matplotlib
sentence-transformers>=2.2.0
faiss-cpu
langchain>=0.1.0
langchain-community
langchain-huggingface
langchain-openai
mlflow>=2.0
pyyaml
loguru
pytest
streamlit
fastapi
uvicorn
transformers
scispacy
# Add specific scispaCy models here if needed, e.g.:
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_sm-0.5.3.tar.gz
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_md-0.5.3.tar.gz
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_lg-0.5.3.tar.gz
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_ner_craft_md-0.5.3.tar.gz
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_ner_jnlpba_md-0.5.3.tar.gz
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_ner_bc5cdr_md-0.5.3.tar.gz
# https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_ner_bionlp13cg_md-0.5.3.tar.gz
